+++ commands/economy/steal.js
@@ -9,13 +9,13 @@ module.exports.run = async (bot, message, args) => {
   if (message.author.id == message.mentions.users.first().id) return;
   embed.setColor('RANDOM');
   embed.setTitle('Cash "Gift"');
-  var result = db.prepare("SELECT coins FROM main WHERE id = ?").get(message.author.id)
-  var result2 = db.prepare("SELECT coins FROM main WHERE id = ?").get(message.mentions.users.first().id)
+  var result = db.prepare("SELECT coins FROM eco WHERE id = ?").get(message.author.id)
+  var result2 = db.prepare("SELECT coins FROM eco WHERE id = ?").get(message.mentions.users.first().id)
   if (result2 === undefined || args[1] > result2.coins) {
     embed.setDescription(`${targetUsername} doesn't have enough money for you to steal. Go find a guy in a lambo ya loser.`)
   } else {
-    db.prepare(`UPDATE main SET coins = ? WHERE id = ?`).run((result.coins - parseInt(args[1])), message.mentions.users.first().id);
-    db.prepare(`UPDATE main SET coins = ? WHERE id = ?`).run((result.coins + parseInt(args[1])), message.author.id);
+    db.prepare(`UPDATE eco SET coins = ? WHERE id = ?`).run((result.coins - parseInt(args[1])), message.mentions.users.first().id);
+    db.prepare(`UPDATE eco SET coins = ? WHERE id = ?`).run((result.coins + parseInt(args[1])), message.author.id);
     embed.setDescription(`${authorUsername} stole ${args[1]} coins from ${targetUsername}`)
   }
   message.channel.send(embed);